refactor: centralize ticket schemas
