+++ src/shared/models/__init__.py
@@ -1,10 +1,10 @@
-from backend.domain.ticket_models import (
+from backend.domain.ticket_status import Impact, Priority, TicketStatus, Urgency
+from backend.schemas.ticket_models import (
+    CleanTicketDTO,
     RawTicketDTO,
     TicketType,
     convert_ticket,
 )
-from backend.domain.ticket_status import Impact, Priority, TicketStatus, Urgency
-from shared.dto import CleanTicketDTO
 
 __all__ = [
     "TicketStatus",
