+++ tests/test_dto.py
@@ -4,7 +4,8 @@ from unittest.mock import AsyncMock
 import pytest
 
 from backend.adapters.mapping_service import MappingService
-from shared.dto import CleanTicketDTO, TicketTranslator
+from backend.schemas.ticket_models import CleanTicketDTO
+from shared.dto import TicketTranslator
 
 
 @pytest.fixture
